requirements.txt
agno>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
apify-client>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
firecrawl-py>=1.0.0

